Made files under src/api/rcos/discord/associations more structured, modified some comments in generate.rs, removed unused crate .
