--- graphql/rcos/discord_assoications/discord_generate.graphql
@@ -1,9 +1,0 @@
-mutation DiscordGenerate($project_id: String!){
-        insert_channel_one(object: {
-        project_id: $project_id
-        channel_id: $channel_id,
-        kind: $kind,
-    }) {
-        channel_id
-    }
-}
