--- src/api/rcos/discord_assoications/account_lookup.rs
@@ -1,24 +1,0 @@
-use crate::api::rcos::{prelude::*, search_strings::resolve_search_string, send_query};
-use crate::error::TelescopeError;
-use crate::api::rcos::discord_assoications::project_info::projects::Variables;
-use crate::api::rcos::discord_assoications::project_info::projects::ResponseData;
-
-/// Projects per page.
-const PER_PAGE: u32 = 20;
-
-/// GraphQL query to get discord account of user.
-#[derive(GraphQLQuery)]
-#[graphql(
-    schema_path = "graphql/rcos/schema.json",
-    query_path = "graphql/rcos/discord_assoications/account_lookup.graphql",
-    response_derives = "Debug,Clone,Serialize"
-)]
-pub struct Account;
-impl Account{
-    pub async fn get(user_id : uuid) -> Result<String, TelescopeError>{
-        send_query::<Self>(Variables {
-            user_id
-        })
-        .await
-    }
-}
