+++ src/api/rcos/discord_assoications/small_group/create_small_group_role.rs
@@ -1,6 +1,5 @@
 //! RCOS API mutation to create a discord channel for a given projct id.
 
-use crate::api::rcos::prelude::*;
 use crate::api::rcos::send_query;
 
 use crate::error::TelescopeError;
@@ -9,20 +8,20 @@ use crate::error::TelescopeError;
 #[derive(GraphQLQuery)]
 #[graphql(
     schema_path = "graphql/rcos/schema.json",
-    query_path = "graphql/rcos/discord_assoications/project/create_project_role.graphql"
+    query_path = "graphql/rcos/discord_assoications/small_group/create_small_group_role.graphql"
 )]
-pub struct CreateOneProjectRole;
+pub struct CreateOneSmallGroupRole;
 
-impl CreateOneProjectRole {
+impl CreateOneSmallGroupRole {
     pub async fn execute(
-        project_id: i64,
+        small_group_id: i64,
         role_id: String,
     ) -> Result<Option<String>, TelescopeError> {
-        send_query::<Self>(create_one_project_role::Variables {
-            project_id,
+        send_query::<Self>(create_one_small_group_role::Variables {
+            small_group_id,
             role_id,
         })
         .await
-        .map(|response| response.insert_project_roles_one.map(|obj| obj.role_id))
+        .map(|response| response.insert_small_group_roles_one.map(|obj| obj.role_id))
     }
 }
